+++ CHANGES.md
@@ -74,6 +74,7 @@ Changes in 9.0
 - Two Lists on same page: export only first, no search in Eligibility/Student.php
 - Remove photos on delete in Student.php & User.php, thank to @jo125ker
 - Remove Student Assignment Submission files on delete in Assignments.php, thank to @khanhchauminh
+- Add microseconds to filename format to make it harder to predict in Assignments.php & StudentAssignments.fnc.php, thanks to @khanhchauminh
 
 Changes in 8.9.5
 ----------------